WIP: Add `formatDateString`
Added `formatDateString` helper in `src/format.ts` to format a string
with date tokens. This helper is a wrapper around `dayjs`'s `format`
method, but inverts the escape token so `[]` is use to indicate what
portions of the string to format as opposed to escape. The reason for
this choice is this helper will be for formatting issue titles and
bodies and it is much simpler to assume that most of the content will
not be date formatted. I.e. 'Complete [YYYY] annual report'.