Fix timezone bug in `calculateNextDate`
Discovered an issue with timezone handling in `calculateNextDate` from
the `ci.yml` workflow. The GitHub actions runners use UTC by default
which caused the "should handle timezones correctly" test to fail, but
not locally. It appears that `RRule` does not handle timezones as well
as what the documentation might suggest, so to address the issue added
handling around the next occurance calculation to convert the provided
date from the given timezone to UTC and then convert the resulting
calculating back to the provided timezone. In addition:

* Formatted the `ci.yml` workflow file and split linting and testing
  into two seperate jobs so the testing job can run on a variety of
  sample timezones.
* Added unit tests around the conversion helpers added to ensure that
  when used together it is possible to recover the exact given input.
* Expanded testing of recurrence rules with an "until" datetime.