Bugfix for start date parsing, v0.1.1
Fixed bug where action would fail on parsing a provided start date in a
configuration file. The error would look something like:

```
Error: Failed to parse configuration: [
  {
    "expected": "date",
    "code": "invalid_type",
    "path": [
      1,
      "start"
    ],
    "message": "Invalid input"
  },
...
```

The root cause is that the configuration file is read in as an object
using `js-yaml`. Zod would then fail seeing that the date was read in as
a string. Had to adjust the `IssueSchema` to coerce the start provided
so it can parse the date from a string. Added an example to the unit
tests that confirm the issue and now works after the fix.