Fix rounding bug to allow for frequent occurrences
Previously rounding for "now" was to midnight. So suppose that there was
an issue with a schedule of "every monday" and now is
2025-10-27T10:00:00 (assume UTC for simplicity). Then rounded now would
be 2025-10-27T00:00:00 and the next Monday would be 2025-11-03T00:00:00
which is in the future and the issue gets skipped. To fix this instead
of just rounding to midnight also subtract 1 second after the rounding.
Added more extensive integration tests to confirm the fix works in
practice and added unit tests for `calculateNextDate` to confirm this
understanding of the calculation.