Fix bug where schedule and due could be the same
Due was erroneously being calculated from "now" instead of "schedule"
which could result in the "due" being the same as "schedule". The
corrected behavior now calculates "due" relative to "schedule". The
practical case where this occurs is when a user specifies schedule and
due to be the same recurrence rule (i.e. "every day"), in this case the
intent is for due to be the following day because otherwise the user
could just omit the due.

Also added more details about this to the `README.md` to clarify this
behavior and prepared for a v0.1.3 release.